Feat(HomePage): home page events lazy loading
